Moved util stuff to helpers directory, fixed .env setup
